get ready for next release
