+++ straxen/plugins/events/multi_scatter.py
@@ -1,8 +1,3 @@
-import numba
-
-from straxen.common import pax_file, get_resource, first_sr1_run, rotate_perp_wires
-from straxen.get_corrections import get_cmt_resource, is_cmt_option
-from straxen.itp_map import InterpolatingMap
 import strax
 import straxen
 import numpy as np
