+++ straxen/plugins/peaks/__init__.py
@@ -33,3 +33,6 @@ from .peaks import *
 
 from . import peaks_per_event
 from .peaks_per_event import *
+
+from . import peaks_corrections
+from .peaks_corrections import *
